+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+## [13.0.3](https://github.com/dhis2/analytics/compare/v13.0.2...v13.0.3) (2021-01-24)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([482b621](https://github.com/dhis2/analytics/commit/482b6211681f723ca28935d0ea6dd6f335d12cb4))
+
+## [13.0.2](https://github.com/dhis2/analytics/compare/v13.0.1...v13.0.2) (2021-01-22)
+
+
+### Bug Fixes
+
+* only pass fixed dimensions as dimension ([#776](https://github.com/dhis2/analytics/issues/776)) ([9c02508](https://github.com/dhis2/analytics/commit/9c02508bb04c1a7db4ee5766828c1b42ac611e63))
+
+## [13.0.1](https://github.com/dhis2/analytics/compare/v13.0.0...v13.0.1) (2021-01-22)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([a1aaba6](https://github.com/dhis2/analytics/commit/a1aaba6b1d4ec25b388c9416f656857a28bb07e7))
+
+# [13.0.0](https://github.com/dhis2/analytics/compare/v12.1.10...v13.0.0) (2021-01-21)
+
+
+### Features
+
+* paging of dynamic dimension items (DHIS2-9274) ([#583](https://github.com/dhis2/analytics/issues/583)) ([670dc79](https://github.com/dhis2/analytics/commit/670dc79a4c55aa070a3219c4d01799bb38b5b631))
+
+
+### BREAKING CHANGES
+
+* DynamicDimension no longer requires context and instead requires dimensionTitle
+
 ## [12.1.10](https://github.com/dhis2/analytics/compare/v12.1.9...v12.1.10) (2021-01-19)